Sign up (not tested)

+++ src/main/java/com/mate/trackq/dao/ProjectDao.java
@@ -8,5 +8,5 @@ public interface ProjectDao {
 
     void create(Project project);
 
-    Project getById(Integer projectId);
+    Project getById(Long projectId);
 }
