+++ src/main/java/com/mate/trackq/service/ProjectService.java
@@ -10,5 +10,5 @@ public interface ProjectService {
 
     void create(Project project);
 
-    Project getById(Integer projectId);
+    Project getById(Long projectId);
 }
