+++ src/main/resources/sql/V001__init.sql
@@ -38,9 +38,9 @@ CREATE TABLE IF NOT EXISTS trackq.user_to_role (
 );
 
 CREATE TABLE IF NOT EXISTS  trackq.user_email_confirmation (
-  id INT PRIMARY KEY AUTO_INCREMENT,
+  id BIGINT PRIMARY KEY AUTO_INCREMENT,
   confirmation_id VARCHAR(255),
-  user_id INT NOT NULL,
+  user_id BIGINT NOT NULL,
   expiration_timestamp timestamp,
   FOREIGN KEY (user_id) REFERENCES users(id)
 );
